requirements.txt
farm-haystack==1.26.4
transformers==4.46.0
sentence-transformers>=2.2.0
haystack-entailment-checker==0.0.4
plotly>=5.0.0
streamlit>=1.30.0
altair>=5.0.0
torch>=2.0.0
faiss-cpu>=1.7.0
python-dotenv
accelerate
